+++ packages/ts-newskit/src/components/slices/lead-article/index.tsx
@@ -43,6 +43,7 @@ export interface LeadArticleProps {
   flag?: string;
   shortSummary?: string;
   contentType?: string;
+  hasVideo: boolean;
   label?: string;
   expirableFlags?: expirableFlagsProps[];
   images?: ImageProps;
@@ -66,10 +67,12 @@ export interface LeadArticleProps {
 
 export const LeadArticle = ({
   article,
-  clickHandler
+  clickHandler,
+  className
 }: {
   article: LeadArticleProps;
   clickHandler: ClickHandlerType;
+  className?: string;
 }) => {
   const {
     id,
@@ -77,6 +80,7 @@ export const LeadArticle = ({
     flag,
     shortSummary,
     contentType,
+    hasVideo,
     images,
     url,
     tag,
@@ -149,12 +153,14 @@ export const LeadArticle = ({
       }}
       columnGap="space040"
       columns={displayArticleVertical || !contentWidth ? '100%' : contentWidth}
+      className={className}
     >
       {hasImage &&
         !hideImage && (
           <Block
             marginBlockEnd={imageTop ? 'space040' : 'space000'}
             marginBlockStart={imageMarginBlockStart}
+            className="lead-image-container"
           >
             <FullWidthCardMediaMob
               {...cardImage}
@@ -192,6 +198,7 @@ export const LeadArticle = ({
           />
         )}
         <ArticleTileInfo
+          hasVideo={hasVideo}
           contentType={contentType}
           expirableFlags={expirableFlags}
           label={label}