+++ packages/ts-newskit/src/slices/content-bucket-3/__tests__/index.test.tsx
@@ -1,14 +1,9 @@
 import React from 'react';
-import { useBreakpointKey } from 'newskit';
-import { render, screen } from '../../../utils/test-utils';
+import { screen } from '../../../utils/test-utils';
 import '@testing-library/jest-dom';
 import { ContentBucket3 } from '..';
 import data from '../../fixtures/data.json';
-
-jest.mock('newskit', () => ({
-  ...jest.requireActual('newskit'),
-  useBreakpointKey: jest.fn().mockReturnValue('xl')
-}));
+import { renderComponent } from '../../../utils';
 
 afterAll(() => {
   jest.clearAllMocks();
@@ -20,67 +15,65 @@ const defaultProps = {
   clickHandler: mockClickHandler
 };
 
-const renderComponent = () => render(<ContentBucket3 {...defaultProps} />);
-
 describe('Render Content Bucket 3 Slice', () => {
   test('Slice matches snapshot', () => {
-    const { asFragment } = renderComponent();
+    const { asFragment } = renderComponent(
+      <ContentBucket3 {...defaultProps} />
+    );
     expect(asFragment()).toMatchSnapshot();
   });
 
   test('Slice matches snapshot for mobile', () => {
-    (useBreakpointKey as any).mockReturnValue('sm');
-    const { asFragment } = renderComponent();
+    const { asFragment } = renderComponent(
+      <ContentBucket3 {...defaultProps} />,
+      'sm'
+    );
     expect(asFragment()).toMatchSnapshot();
   });
 });
 
 describe('Content Bucket 3 Articles list ', () => {
-  test('articleBorder render after odd numbered items', () => {
-    (useBreakpointKey as any).mockReturnValue('xl');
-    renderComponent();
-    const articleContainer = screen.getByTestId('article-container');
+  test("<ArticleDividerXL> renders in article-container at 'xl' breakpoint", () => {
+    renderComponent(<ContentBucket3 {...defaultProps} />, 'xl');
+    const articleContainer = screen.getByTestId('article-container-desktop');
+    const articleBorder = articleContainer.querySelector(
+      '[data-testid="article-stack-large-divider"]'
+    );
 
-    const articleItem4 = articleContainer.lastElementChild!
-      .previousElementSibling;
-    const articleHR = articleContainer.lastElementChild;
-
-    expect(articleHR).toHaveAttribute('data-testid', 'divider');
-    expect(articleItem4!.getElementsByTagName('hr').length).toBe(2);
+    expect(articleBorder).toBeInTheDocument();
   });
 
   test("articleTopBorder renders correctly at 'xl' breakpoint", () => {
-    (useBreakpointKey as any).mockReturnValue('xl');
-    renderComponent();
-    const articleContainer = screen.getByTestId('article-container');
-
+    renderComponent(<ContentBucket3 {...defaultProps} />, 'xl');
+    const articleContainer = screen.getByTestId('article-container-desktop');
     const articleItem1 = articleContainer.firstElementChild;
-    const articleItem4 = articleContainer.lastElementChild!
+    const articleItem4 = articleContainer!.lastElementChild!
       .previousElementSibling;
 
-    expect(articleItem1!.getElementsByTagName('hr').length).toBe(1);
-    expect(articleItem4!.getElementsByTagName('hr').length).toBe(2);
+    expect(articleItem1!.getElementsByTagName('hr').length).toBe(2);
+    expect(articleItem4!.getElementsByTagName('hr').length).toBe(3);
   });
 
   test("articleTopBorder renders correctly below 'lg' breakpoint", () => {
-    (useBreakpointKey as any).mockReturnValue('md');
-    renderComponent();
-    const articleContainer = screen.getByTestId('article-container');
-
+    renderComponent(<ContentBucket3 {...defaultProps} />, 'lg');
+    const articleContainer = screen.getByTestId('article-container-desktop');
     const articleItem1 = articleContainer.firstElementChild;
+    const articleBorder = articleContainer.querySelector(
+      '[data-testid="article-stack-large-divider"]'
+    );
 
+    expect(articleBorder).not.toBeVisible();
     expect(articleItem1!.nextSibling).toHaveAttribute('data-testid', 'divider');
   });
 
   test("articleTopBorder renders correctly at 'lg' breakpoint", () => {
-    (useBreakpointKey as any).mockReturnValue('lg');
-    renderComponent();
-    const articleContainer = screen.getByTestId('article-container');
+    renderComponent(<ContentBucket3 {...defaultProps} />, 'lg');
+    const articleContainer = screen.getByTestId('article-container-desktop');
+    const articleItem1 = articleContainer!.firstElementChild;
+    const articleItem4 = articleContainer!.lastElementChild!
+      .previousElementSibling;
 
-    const articleItem1 = articleContainer.firstElementChild;
-    const articleItem4 = articleContainer.lastElementChild;
-
-    expect(articleItem1!.getElementsByTagName('hr').length).toBe(1);
-    expect(articleItem4!.getElementsByTagName('hr').length).toBe(2);
+    expect(articleItem1!.getElementsByTagName('hr').length).toBe(2);
+    expect(articleItem4!.getElementsByTagName('hr').length).toBe(3);
   });
 });