+++ packages/ts-newskit/src/utils/__tests__/clear-credits-and-caption.test.tsx
@@ -6,6 +6,7 @@ import {
 describe('clearCreditsAndCaptionObject', () => {
   it('should clear credits and caption of a single ArticleProps object', () => {
     const leadArticle: ArticleProps = {
+      hasVideo: false,
       id: '1234',
       headline: 'Example Headline',
       url: '',