+++ packages/article-in-depth/CHANGELOG.md
@@ -3,6 +3,86 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.76.79](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.76.78...@times-components/article-in-depth@3.76.79) (2023-10-27)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+## [3.76.78](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.76.77...@times-components/article-in-depth@3.76.78) (2023-10-27)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+## [3.76.77](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.76.76...@times-components/article-in-depth@3.76.77) (2023-10-26)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+## [3.76.76](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.76.75...@times-components/article-in-depth@3.76.76) (2023-10-26)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+## [3.76.75](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.76.74...@times-components/article-in-depth@3.76.75) (2023-10-25)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+## [3.76.74](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.76.73...@times-components/article-in-depth@3.76.74) (2023-10-25)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+## [3.76.73](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.76.72...@times-components/article-in-depth@3.76.73) (2023-10-24)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+## [3.76.72](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.76.71...@times-components/article-in-depth@3.76.72) (2023-10-24)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+## [3.76.71](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.76.70...@times-components/article-in-depth@3.76.71) (2023-10-23)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+## [3.76.70](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.76.69...@times-components/article-in-depth@3.76.70) (2023-10-23)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
 ## [3.76.69](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.76.68...@times-components/article-in-depth@3.76.69) (2023-10-19)
 
 **Note:** Version bump only for package @times-components/article-in-depth