+++ packages/article-magazine-standard/CHANGELOG.md
@@ -3,6 +3,86 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.72.79](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.72.78...@times-components/article-magazine-standard@3.72.79) (2023-10-27)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
+## [3.72.78](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.72.77...@times-components/article-magazine-standard@3.72.78) (2023-10-27)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
+## [3.72.77](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.72.76...@times-components/article-magazine-standard@3.72.77) (2023-10-26)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
+## [3.72.76](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.72.75...@times-components/article-magazine-standard@3.72.76) (2023-10-26)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
+## [3.72.75](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.72.74...@times-components/article-magazine-standard@3.72.75) (2023-10-25)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
+## [3.72.74](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.72.73...@times-components/article-magazine-standard@3.72.74) (2023-10-25)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
+## [3.72.73](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.72.72...@times-components/article-magazine-standard@3.72.73) (2023-10-24)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
+## [3.72.72](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.72.71...@times-components/article-magazine-standard@3.72.72) (2023-10-24)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
+## [3.72.71](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.72.70...@times-components/article-magazine-standard@3.72.71) (2023-10-23)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
+## [3.72.70](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.72.69...@times-components/article-magazine-standard@3.72.70) (2023-10-23)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
 ## [3.72.69](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.72.68...@times-components/article-magazine-standard@3.72.69) (2023-10-19)
 
 **Note:** Version bump only for package @times-components/article-magazine-standard